requirements.txt
aiogram==2.14
fastapi==0.109.2
uvicorn==0.27.1
sqlalchemy==2.0.27
python-jose==3.3.0
python-multipart==0.0.6
passlib==1.7.4
bcrypt==4.1.2
python-dotenv==1.0.1
aiosqlite==0.19.0
jinja2==3.1.3
python-socketio==5.11.1 